Add one line to check if we need to convert tree to workspace

+++ testRun.sh
@@ -38,7 +38,7 @@ then
   rm Selections_Run.C
   rm DataSelections_Run.C
 else
-  echo "Donot do selections,just copy tree file "
+  echo "Do not apply any selections ,just copy tree file "
   cp ${TreePath}${procs}_node_${node}_${year}.root ./${procs}_node_${node}_${year}.root
   cp ${DataTreeFile} ./Data_13TeV_${cat}_${year}.root
 fi
@@ -63,10 +63,12 @@ fi
 python trees2ws.py --inputConfig HHWWgg_config.py --inputTreeFile ./${procs}_node_${node}_${year}.root --inputMass node_${node} --productionMode ${procs}  --year ${year} --doSystematics
 
 # data tree to data ws
-python trees2ws_data.py --inputConfig HHWWgg_config.py --inputTreeFile ./Data_13TeV_${cat}_${year}.root 
-
+if [ ! -f "../Background/Input/${year}/allData.root" ]; then
+  echo "Do not have an input Data WS, convert tree to workspace "
+  python trees2ws_data.py --inputConfig HHWWgg_config.py --inputTreeFile ./Data_13TeV_${cat}_${year}.root 
+  mv ws/Data_13TeV_${cat}_${year}.root ../Background/Input/${year}/allData.root
+fi
 mv ws_${procs}/GluGluToHHTo2G2l2nu_node_${node}_${year}_${procs}.root ../Signal/Input/output_M125_${procs}_node_${node}_${cat}.root
-mv ws/Data_13TeV_${cat}_${year}.root ../Background/Input/${year}/allData.root
 rm ${procs}_node_${node}_${year}.root
 rm Data_13TeV_${cat}_${year}.root
